simple_trader/projects/nasdaq_ipo/requirements.txt
yfinance>=0.2.18
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
pytest>=7.0.0
mypy>=1.0.0
ruff>=0.0.250
